Image_Expression_Parser: fix dependency version
This version 2.4.1 depends on jama:Jama:1.0.2, when the artifact is
actually available as jama:jama:1.0.2. Not sure exactly how that
happened, but updating to the latest SNAPSHOT version fixes the build.

This is probably what we want anyway, so that the entire multi-module
reactor stays consistent. Otherwise, it is rather unintuitive for the
Image_Expression_Parser to depend on jep, but not the version of jep
actually being built during that same build!