Switch to ImgLib2
Ever since ImgLib2 was announced, ImgLib1 was kind of a zombie: a dead man
still walking, confusing new developers as to which version they should
use, and using up valuable time from maintainers who somehow had to manage
twice as many dependencies as before.

To remedy this, ImgLib1 was deprecated over a year ago. It is high time to
move away from it.

This here monster commit ports the Image Expression Parser to ImgLib2.

The Normalizer had to be reimplemented because it got lost in the broken
ImgLib1->ImgLib2 conversions that got removed somewhere along the lines
(the Normalize class in algorithms performs a *different* normalization).

The Floyd-Steinberg dithering in ImgLib2 Algorithms does nothing at all at
the time of writing because the code is commented out (probably in an
attempt to keep this class around until the time the original author could
fix it). Therefore, the corresponding unit test unfortunately had to be
commented out. To help existing users with figuring out what went wrong, a
helpful exception was added in case the dithering did nothing. That way,
once ImgLib2 Algorithms is fixed, everything will magically start working
again.

As the complete API is upgraded to the (backwards-incompatible) ImgLib2
data structures, Semantic Versioning dictates that we bump to the next
major version.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>