Replace deprecated equalIterationOrder() call
It would have been quite nice if the commit removing the
equalIterationOrder() method, or at least the commit deprecating it, let
alone the Javadoc of the deprecated method, have mentioned that the Util
class has the suggested replacement.

But half an hour of research accompanied by nine and a half years of Git
expertise could not stop me from uncovering this secret! Hah! Take this!
Pow! Bam! Smack! I win!

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>