Bump to preemptively-bumped parent
We bumped the version of ImageJ Expression Parser preemptively to be
able to switch to completely reproducible builds.

Unfortunately, this bump is quite bumpy: just in time for the final
release, after several years of beta phase, the names of the algorithm
packages were changed, and they contain the same hyper sphere
neighborhood class, so we have to configure quite a bit to make that
work.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>